refactor: move thread runtime pool registry
